reorganize empty messages and enhance ListItems component logic
