+++ frontend/src/data/emptyMessages.js
@@ -1,0 +1,14 @@
+const emptyMessages = {
+  'my-recipes':
+    'Nothing has been added to your recipes list yet. Please browse our recipes and add your favorites for easy access in the future.',
+  'my-favorites':
+    'Nothing has been added to your favorite recipes list yet. Please browse our recipes and add your favorites for easy access in the future.',
+  recipes:
+    'Nothing has been added to the recipes list yet. Please browse our recipes and add your favorites for easy access in the future.',
+  followers:
+    'There are currently no followers on your account. Please engage our visitors with interesting content and draw their attention to your profile.',
+  following:
+    'Your account currently has no subscriptions to other users. Learn more about our users and select those whose content interests you.',
+};
+
+export default emptyMessages;
